BREAK: modernizing code and api
discontinues support for node < 8
uses `readArrayObj` for previous ReadArray.obj
uses `readArrayObj` for previous ReadArray.obj
exports `through` and `throughObj` for static `Through` functions
