--- lib/readarray.js
@@ -1,73 +1,0 @@
-/**
- * @copyright 2014- commenthol
- * @licence MIT
- */
-
-'use strict'
-
-const util = require('util')
-const Readable = require('stream').Readable
-
-/**
- * Read from an Array and push into stream.
- *
- * Takes care on pausing to push to the stream if pipe is saturated.
- *
- * @constructor
- * @param {Object} [options] - Stream Options `{encoding, highWaterMark, objectMode, ...}`
- * @param {Array} array - array to push down as stream (If array is an array of objects set `objectMode:true` or use `ReadArray.obj`)
- * @return {Readable} A readable stream
- */
-function ReadArray (options, array) {
-  if (!(this instanceof ReadArray)) {
-    return new ReadArray(options, array)
-  }
-
-  if (Array.isArray(options)) {
-    array = options
-    options = {}
-  }
-
-  options = Object.assign({}, options)
-  Readable.call(this, options)
-
-  this._array = array || []
-  this._cnt = this._array.length
-
-  return this
-}
-
-util.inherits(ReadArray, Readable)
-
-/**
- * @private
- * @param {Number} size
- */
-ReadArray.prototype._read = function (size) {
-  while ((this._cnt--) > 0) {
-    // istanbul ignore if
-    if (!this.push(this._array.shift())) {
-      return
-    }
-  }
-  this.push(null) // end of stream
-}
-
-/**
- * Shortcut for ObjectMode
- *
- * @param {Object} [options] - Stream Options `{encoding, highWaterMark, decodeStrings, ...}`
- * @param {Array} array - array to push down as stream
- * @return {Readable} A readable stream
- */
-ReadArray.obj = function (options, array) {
-  // istanbul ignore else
-  if (Array.isArray(options)) {
-    array = options
-    options = {}
-  }
-  options = Object.assign({}, options, { objectMode: true })
-  return new ReadArray(options, array)
-}
-
-module.exports = ReadArray
