+++ package.json
@@ -33,6 +33,10 @@
     "test": "mocha"
   },
   "eslintConfig": {
+    "env": {
+      "node": true,
+      "mocha": true
+    },
     "extends": "standard",
     "plugins": [
       "standard"
@@ -42,8 +46,7 @@
     }
   },
   "dependencies": {
-    "lodash.omit": "^4.5.0",
-    "streamss-through": "^1.0.2"
+    "streamss-through": "^2.0.0-0"
   },
   "devDependencies": {
     "eslint": "^6.6.0",
